Ya compila

+++ s2_centrodeportivo-back/src/main/java/co/edu/uniandes/csw/centrodeportivo/entities/RutinaEntity.java
@@ -14,15 +14,17 @@ import javax.persistence.Entity;
  */
 @Entity
 public class RutinaEntity extends BaseEntity implements Serializable {
-     private int identificadorRutina;
+    
+   
+    private Long identificadorRutina;
     private String nombre;
     private boolean estadoTerminado;
 
-    public int getIdentificadorRutina() {
+    public long getIdentificadorRutina() {
         return identificadorRutina;
     }
 
-    public void setIdentificadorRutina(int identificadorRutina) {
+    public void setIdentificadorRutina(Long identificadorRutina) {
         this.identificadorRutina = identificadorRutina;
     }
